Add cli support to repository
