+++ controller/xpg_deletion_protection.go
@@ -2,7 +2,7 @@ package controller
 
 import (
 	"context"
-	"github.com/crossplane-contrib/provider-kubernetes/apis/object/v1alpha1"
+	xkube "github.com/crossplane-contrib/provider-kubernetes/apis/object/v1alpha1"
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 	logging "sigs.k8s.io/controller-runtime/pkg/log"
 	"time"
@@ -15,20 +15,12 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-type jsonpatch1 struct {
-	Op    jsonOp `json:"op,omitempty"`
-	Path  string `json:"path,omitempty"`
-	Value int    `json:"value,-"`
-}
-
-var stackgresManagedKey = "stackgres.io~1managed-by-server-side-apply"
-
-type xpostgreSQLDeletionProtectionReconciler struct {
+type XPostgreSQLDeletionProtectionReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
 }
 
-func (p *xpostgreSQLDeletionProtectionReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+func (p *XPostgreSQLDeletionProtectionReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	ctx = context.WithValue(ctx, "now", time.Now())
 
 	log := logging.FromContext(ctx, "namespace", req.Namespace, "instance", req.Name)
@@ -47,14 +39,14 @@ func (p *xpostgreSQLDeletionProtectionReconciler) Reconcile(ctx context.Context,
 	}
 
 	if inst.DeletionTimestamp != nil {
-		log.Info("Scaling down the postgres instance to 0")
+		log.Info("Deleting database")
 		err = p.deletingPostgresDB(ctx, inst)
 	}
 
 	return ctrl.Result{RequeueAfter: requeueTime}, err
 }
 
-func (p *xpostgreSQLDeletionProtectionReconciler) handleDeletionProtection(ctx context.Context, inst *vshnv1.XVSHNPostgreSQL) error {
+func (p *XPostgreSQLDeletionProtectionReconciler) handleDeletionProtection(ctx context.Context, inst *vshnv1.XVSHNPostgreSQL) error {
 	protectionEnabled := inst.Spec.Parameters.Backup.DeletionProtection
 	retention := inst.Spec.Parameters.Backup.DeletionRetention
 	baseObj := &vshnv1.XVSHNPostgreSQL{
@@ -79,58 +71,21 @@ func (p *xpostgreSQLDeletionProtectionReconciler) handleDeletionProtection(ctx c
 	return nil
 }
 
-func (p *xpostgreSQLDeletionProtectionReconciler) deletingPostgresDB(ctx context.Context, inst *vshnv1.XVSHNPostgreSQL) error {
+func (p *XPostgreSQLDeletionProtectionReconciler) deletingPostgresDB(ctx context.Context, inst *vshnv1.XVSHNPostgreSQL) error {
 	log := logging.FromContext(ctx, "namespace", inst.GetNamespace(), "instance", inst.GetName())
 
-	log.V(1).Info("Getting postgres statefulset")
-	o := &v1alpha1.Object{
+	log.V(1).Info("Deleting sgcluster object")
+	o := &xkube.Object{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: inst.Name + "-cluster",
 		},
 	}
-	err := p.Delete(ctx, o)
-	/*
 
-		k := types.NamespacedName{
-			Namespace: inst.Status.InstanceNamespace,
-			Name:      inst.Name,
-		}
-
-
-		err := p.Get(ctx, k, ss)
-		if err != nil {
-			return fmt.Errorf("cannot get postgres statefulset %v", err)
-		}
-
-			var r int32 = 0
-			ss.Spec.Replicas = &r
-			// setting annotation to false does not work
-			ss.Annotations = nil
-			ss.Labels = nil
-			ss.OwnerReferences = nil
-
-
-
-		log.V(1).Info("Updating postgres statefulset")
-		pa := []jsonpatch1{
-			{
-				Op:    opReplace,
-				Path:  "/spec/replicas",
-				Value: 0,
-			},
-		}
-		a, err := json.Marshal(pa)
-		b := client.RawPatch(types.JSONPatchType, a)
-		err = p.Delete(ctx, ss, b)
-		if err != nil {
-			return fmt.Errorf("cannot update statefulset %v", err)
-		}
-	*/
-	return err
+	return p.Delete(ctx, o)
 }
 
 // SetupWithManager sets up the controller with the Manager.
-func (p *xpostgreSQLDeletionProtectionReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (p *XPostgreSQLDeletionProtectionReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&vshnv1.XVSHNPostgreSQL{}).
 		Complete(p)
